feat(parser): fill meta for each entity at each parsing step
