+++ internal/compiler/parser/manifest.go
@@ -6,6 +6,7 @@ import (
 	yaml "gopkg.in/yaml.v3"
 
 	src "github.com/nevalang/neva/internal/compiler/sourcecode"
+	"github.com/nevalang/neva/internal/compiler/sourcecode/core"
 )
 
 func (p Parser) ParseManifest(raw []byte) (src.ModuleManifest, error) {
@@ -17,14 +18,14 @@ func (p Parser) ParseManifest(raw []byte) (src.ModuleManifest, error) {
 }
 
 func processParsedManifest(manifest src.ModuleManifest) src.ModuleManifest {
-	deps := make(map[string]src.ModuleRef, len(manifest.Deps))
+	deps := make(map[string]core.ModuleRef, len(manifest.Deps))
 
 	for alias, dep := range manifest.Deps {
 		path := dep.Path
 		if path == "" {
 			path = alias
 		}
-		deps[alias] = src.ModuleRef{
+		deps[alias] = core.ModuleRef{
 			Path:    path,
 			Version: dep.Version,
 		}
